refactor(example): refactor structure of the example app to allow for more examples
