+++ .yarnrc.yml
@@ -1,3 +1,5 @@
 enableImmutableInstalls: false
 
 nodeLinker: node-modules
+
+nmHoistingLimits: workspaces
