+++ apps/common-app/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "common-app",
+  "version": "0.0.1",
+  "private": true,
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "peerDependencies": {
+    "react": "*",
+    "react-native": "*",
+    "react-native-audio-browser": "*"
+  },
+  "dependencies": {
+    "@react-native-vector-icons/fontawesome6": "^12.0.1"
+  },
+  "devDependencies": {
+    "react": "*",
+    "react-native": "*"
+  }
+}
