+++ apps/example-native/.prettierrc.js
