+++ apps/example-native/.watchmanconfig
