+++ apps/example-native/android/app/proguard-rules.pro
