+++ apps/example-native/android/app/src/main/AndroidManifest.xml
