+++ apps/example-native/android/app/src/main/res/drawable/rn_edit_text_material.xml
