+++ apps/example-native/android/app/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
