+++ apps/example-native/android/app/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
