+++ apps/example-native/android/build.gradle
