+++ apps/example-native/android/gradle.properties
