+++ apps/example-native/ios/AudioBrowserExample.xcodeproj/xcshareddata/xcschemes/AudioBrowserExample.xcscheme
