+++ apps/example-native/ios/AudioBrowserExample/Images.xcassets/AppIcon.appiconset/Contents.json
