+++ apps/example-native/react-native.config.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const pkg = require('../package.json')
+const pkg = require('../../package.json')
 
 /**
  * @type {import('@react-native-community/cli-types').Config}
@@ -12,7 +12,7 @@ module.exports = {
   },
   dependencies: {
     [pkg.name]: {
-      root: path.join(__dirname, '..'),
+      root: path.join(__dirname, '..', '..'),
     },
   },
 }
