+++ apps/example-native/src/hooks/useBrowserHistory.ts
