+++ apps/example-native/tsconfig.json
@@ -6,7 +6,7 @@
     "strict": true,
     "baseUrl": ".",
     "paths": {
-      "react-native-audio-browser": ["../src"]
+      "react-native-audio-browser": ["../../src"]
     }
   }
 }
