--- example/metro.config.js
@@ -1,15 +1,0 @@
-const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
-const path = require('path');
-const root = path.resolve(__dirname, '..');
-
-/**
- * Metro configuration
- * https://facebook.github.io/metro/docs/configuration
- *
- * @type {import('metro-config').MetroConfig}
- */
-const config = {
-  watchFolders: [root],
-};
-
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
