+++ package.json
@@ -55,7 +55,7 @@
     "README.md"
   ],
   "workspaces": [
-    "example"
+    "apps/*"
   ],
   "repository": "https://github.com/puckey/react-native-audio-browser.git",
   "author": "puckey",
